move test files
